+++ README.md
@@ -1,4 +1,4 @@
-# gboost_module
+# gbnet
 
 Gradient Boosting Modules for pytorch
 
@@ -6,14 +6,15 @@ Gradient Boosting Modules for pytorch
 
 Gradient Boosting Machines only require gradients and, for modern packages, hessians to train. Pytorch (and other neural network packages) calculates gradients and hessians. GBMs can therefore be fit as the first layer in neural networks using Pytorch. This package provides access to XGBoost and LightGBM as Pytorch Modules to do exactly this.
 
-CatBoost is supported in an experimental capacity since the current gboost_module integration with CatBoost is not as performant as the other GBDT packages.
+CatBoost is supported in an experimental capacity since the current gbnet integration with CatBoost is not as performant as the other GBDT packages.
 
 ## Install
 
-`pip install gboost_module`
+`pip install gbnet`
 
 ## Troubleshooting
-1. Currently, the biggest difference between training using `gboost_module` vs basic `torch`, is that `gboost_module`, like basic usage of `xgboost` and `lightgbm`, requires the entire dataset to be fed in. Cached predictions allow these packages to train quickly, and caching cannot happen if input batches change with each training/boosting round. Some additional info is provided in [#12](https://github.com/mthorrell/gboost_module/issues/12).
+
+1. Currently, the biggest difference between training using `gbnet` vs basic `torch`, is that `gbnet`, like basic usage of `xgboost` and `lightgbm`, requires the entire dataset to be fed in. Cached predictions allow these packages to train quickly, and caching cannot happen if input batches change with each training/boosting round. Some additional info is provided in [#12](https://github.com/mthorrell/gbnet/issues/12).
 
 ## Basic training of a GBM for comparison to existing packages
 
@@ -25,7 +26,7 @@ import numpy as np
 import xgboost as xgb
 import torch
 
-from gboost_module import lgbmodule, xgbmodule
+from gbnet import lgbmodule, xgbmodule
 
 # Generate Dataset
 np.random.seed(100)
@@ -99,7 +100,7 @@ import time
 import numpy as np
 import torch
 
-from gboost_module import lgbmodule, xgbmodule
+from gbnet import lgbmodule, xgbmodule
 
 
 # Create new module that jointly trains multi-output xgboost and lightgbm models
@@ -143,7 +144,7 @@ for i in range(100):
     preds = gbp(X)
 
     loss = mse(preds, torch.Tensor(Y))
-    loss.backward(create_graph=True)  # create_graph=True required for any gboost_module
+    loss.backward(create_graph=True)  # create_graph=True required for any gbnet
     losses.append(loss.detach().numpy().copy())
 
     gbp.gb_step(X)  # required to update the gbms
@@ -151,4 +152,5 @@ for i in range(100):
 t1 = time.time()
 print(t1 - t0)  # 5.821
 ```
+
 <img width="750" alt="image" src="https://github.com/mthorrell/gbmodule/assets/15166269/949c7000-7fc3-4600-8916-03cdf60eeeb8">
