+++ gbnet/xgbmodule.py
@@ -1,5 +1,3 @@
-from typing import Tuple
-
 import numpy as np
 import torch
 from torch import nn
