+++ setup.py
@@ -1,14 +1,14 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="gboost_module",
-    version="0.1.4",
+    name="gbnet",
+    version="0.1.5",
     author="Michael Horrell",
     author_email="mthorrell@github.com",
-    description="Torch modules using popular boosting libraries",
+    description="Gradient boosting libraries integrated with pytorch",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
-    url="https://github.com/mthorrell/gboost_module",
+    url="https://github.com/mthorrell/gbnet",
     packages=find_packages(),
     install_requires=[
         "numpy",
